restructure to make more reusable
