+++ test/actOnDispute.js.test
@@ -1,0 +1,18 @@
+import { processDisputes } from '../src/disputes/actOnOpenDisputes'
+
+describe('actOnDispute', () => {
+  const TxController = {
+  }
+
+  const _asyncMockResponse = async response => {
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        resolve(response)
+      }, 1000)
+    })
+  }
+
+  beforeEach(async () => {
+
+  })
+})
