--- test/bot.js.test
@@ -1,3 +1,0 @@
-describe('Bot', () => {
-  
-})
